Move chunk_{128,256}bit_perm_idx_lut.h, chunk_permute_table.h to arch/shared.
