+++ arch/arm/chunkset_neon.c
@@ -8,7 +8,7 @@
 #include "zsanitizer.h"
 #include "zmemory.h"
 #include "neon_intrins.h"
-#include "arch/generic/chunk_128bit_perm_idx_lut.h"
+#include "arch/shared/chunk_128bit_perm_idx_lut.h"
 
 typedef uint8x16_t chunk_t;
