+++ arch/loongarch/chunkset_lasx.c
@@ -13,7 +13,7 @@
 #include "lasxintrin_ext.h"
 #include "lsxintrin_ext.h"
 
-#include "arch/generic/chunk_256bit_perm_idx_lut.h"
+#include "arch/shared/chunk_256bit_perm_idx_lut.h"
 
 typedef __m256i chunk_t;
 typedef __m128i halfchunk_t;
