+++ arch/loongarch/chunkset_lsx.c
@@ -11,7 +11,7 @@
 
 #include <lsxintrin.h>
 #include "lsxintrin_ext.h"
-#include "arch/generic/chunk_128bit_perm_idx_lut.h"
+#include "arch/shared/chunk_128bit_perm_idx_lut.h"
 
 typedef __m128i chunk_t;
