+++ arch/shared/chunk_128bit_perm_idx_lut.h
