+++ arch/shared/chunk_256bit_perm_idx_lut.h
