+++ arch/x86/chunkset_avx512.c
@@ -7,7 +7,7 @@
 #include "zbuild.h"
 #include "zmemory.h"
 
-#include "arch/generic/chunk_256bit_perm_idx_lut.h"
+#include "arch/shared/chunk_256bit_perm_idx_lut.h"
 #include <immintrin.h>
 #include "x86_intrins.h"
