+++ arch/x86/chunkset_ssse3.c
@@ -9,7 +9,7 @@
 #include "zmemory.h"
 
 #include <immintrin.h>
-#include "arch/generic/chunk_128bit_perm_idx_lut.h"
+#include "arch/shared/chunk_128bit_perm_idx_lut.h"
 
 typedef __m128i chunk_t;
